Add WIP refactor of CSV creation to CSVHelperFunctions.groovy
